Migrated file transfer to new data flow
