+++ app/data/file-explorer/data-sources/FileExplorerLegacyDataSource.js
@@ -2,18 +2,22 @@ import path from 'path';
 import { Promise } from 'bluebird';
 import moment from 'moment';
 import findLodash from 'lodash/find';
-import { exec } from 'child_process';
+import { exec, spawn } from 'child_process';
 import { log } from '../../../utils/log';
 import {
   isArray,
   isEmpty,
+  niceBytes,
+  percentage,
   splitIntoLines,
   undefinedOrNull,
 } from '../../../utils/funcs';
 import { DEVICES_LABEL } from '../../../constants';
 import { DEVICE_TYPE } from '../../../enums';
-import { getExtension } from '../../../utils/files';
+import { baseName, getExtension } from '../../../utils/files';
 import { mtp as _mtpCli } from '../../../utils/binaries';
+import { msToTime, unixTimestampNow } from '../../../utils/date';
+import { checkIf } from '../../../utils/checkIf';
 
 export class FileExplorerLegacyDataSource {
   constructor() {
@@ -160,6 +164,191 @@ export class FileExplorerLegacyDataSource {
     );
 
     return !stderr;
+  }
+
+  /**
+   * description - file transfer helper
+   *
+   * @param {[string]} cmdArgs - command line strings in an array
+   * @param {errorCallback} onError
+   * @param {progressCallback} onProgress
+   * @param {completedCallback} onCompleted
+   * @return {Promise<{data: boolean, error: string|null, stderr: string|null}>}
+   * @private
+   */
+  async _transferFiles({ cmdArgs, onError, onProgress, onCompleted }) {
+    try {
+      checkIf(cmdArgs, 'array');
+      checkIf(onError, 'function');
+      checkIf(onProgress, 'function');
+      checkIf(onCompleted, 'function');
+
+      const handletransferListTimeInterval = 1000;
+      let transferList = {};
+      let prevCopiedBlockSize = 0;
+      let currentCopiedBlockSize = 0;
+      let startTime = 0;
+      let prevCopiedTime = 0;
+      let currentCopiedTime = 0;
+      let bufferedOutput = null;
+
+      let handleTransferListInterval = setInterval(() => {
+        if (transferList === null) {
+          clearInterval(handleTransferListInterval);
+          handleTransferListInterval = 0;
+          return null;
+        }
+
+        if (Object.keys(transferList).length < 1) {
+          return null;
+        }
+
+        const {
+          percentage: _percentage,
+          currentFile,
+          activeFileSize,
+          activeFileSent,
+        } = transferList;
+
+        const copiedTimeDiff = currentCopiedTime - prevCopiedTime;
+        const speed =
+          prevCopiedBlockSize &&
+          prevCopiedBlockSize - currentCopiedBlockSize > 0
+            ? (prevCopiedBlockSize - currentCopiedBlockSize) *
+              (1000 / copiedTimeDiff)
+            : 0;
+        const _speed = speed ? `${niceBytes(speed)}` : `--`;
+        const elapsedTime = msToTime(currentCopiedTime - startTime);
+        prevCopiedTime = currentCopiedTime;
+        prevCopiedBlockSize = currentCopiedBlockSize;
+
+        onProgress({
+          elapsedTime,
+          speed: _speed,
+          percentage: _percentage,
+          currentFile,
+          activeFileSize,
+          activeFileSent,
+        });
+      }, handletransferListTimeInterval);
+
+      const cmd = spawn(this.mtpCli, [...cmdArgs], {
+        shell: true,
+      });
+
+      cmd.stdout.on('data', (data) => {
+        bufferedOutput = data.toString();
+
+        if (startTime === 0) {
+          startTime = unixTimestampNow();
+        }
+
+        if (
+          typeof bufferedOutput === 'undefined' ||
+          bufferedOutput === null ||
+          bufferedOutput.length < 1
+        ) {
+          return null;
+        }
+
+        const _bufferedOutput = splitIntoLines(bufferedOutput).filter(
+          (a, index) => !this._filterOutMtpLines(a, index)
+        );
+
+        if (_bufferedOutput.length < 1) {
+          return null;
+        }
+
+        for (let i = 0; i < _bufferedOutput.length; i += 1) {
+          const item = _bufferedOutput[i];
+          const bufferedOutputSplit = item.split(' ');
+
+          if (bufferedOutputSplit.length < 1) {
+            return null;
+          }
+
+          const totalLength = bufferedOutputSplit.length;
+          const eventIndex = 0;
+          const filePathStartIndex = 1;
+          const filePathEndIndex = totalLength - 3;
+          const currentProgressSizeIndex = totalLength - 2;
+          const totalFileSizeIndex = totalLength - 1;
+
+          const event = bufferedOutputSplit[eventIndex];
+          const matchedItem = item.match(/(\d+?\d*)\s(\d+?\d*)$/);
+          if (matchedItem === null) {
+            return null;
+          }
+
+          const matchedItemSplit = matchedItem[0].split(' ');
+          const currentProgressSize = parseInt(matchedItemSplit[0], 10);
+          const totalFileSize = parseInt(matchedItemSplit[1], 10);
+
+          if (event === `:done`) {
+            prevCopiedBlockSize = 0;
+            currentCopiedBlockSize = 0;
+            prevCopiedTime = 0;
+            currentCopiedTime = 0;
+            return null;
+          }
+
+          if (
+            totalLength < 3 ||
+            event !== `:progress` ||
+            currentProgressSizeIndex < 2 ||
+            totalFileSizeIndex < 3
+          ) {
+            return null;
+          }
+
+          const filePath = bufferedOutputSplit
+            .slice(filePathStartIndex, filePathEndIndex + 1)
+            .join(' ');
+
+          const perc = percentage(currentProgressSize, totalFileSize);
+          currentCopiedBlockSize = totalFileSize - currentProgressSize;
+          currentCopiedTime = unixTimestampNow();
+
+          transferList = {
+            currentFile: filePath,
+            activeFileSize: totalFileSize,
+            activeFileSent: currentProgressSize,
+            percentage: perc,
+            currentCopiedBlockSize,
+            currentCopiedTime,
+          };
+        }
+      });
+
+      // capture error
+      cmd.stderr.on('data', (error) => {
+        const { filteredError } = this._cleanJunkMtpError({ error });
+
+        if (undefinedOrNull(filteredError) || filteredError.length < 1) {
+          return null;
+        }
+
+        onError({
+          error,
+          stderr: null,
+          data: null,
+        });
+
+        transferList = null;
+      });
+
+      cmd.on('exit', () => {
+        transferList = null;
+
+        onCompleted();
+      });
+
+      return { error: null, stderr: null, data: true };
+    } catch (e) {
+      log.error(e);
+
+      return { error: null, stderr: null, data: false };
+    }
   }
 
   /**
@@ -482,4 +671,133 @@ export class FileExplorerLegacyDataSource {
       return false;
     }
   }
+
+  /**
+   * @typedef {function(errorCallbackInfo)} errorCallback
+   * @callback errorCallback
+   * @param {errorCallbackInfo} args - error object
+   */
+
+  /**
+   * @typedef {Object} errorCallbackInfo
+   * @property {string} error - error text
+   * @property {string} stderr - std error text
+   * @property {null} data - data
+   */
+
+  /**
+   * @typedef {function(progressCallbackInfo)} progressCallback
+   * @callback progressCallback
+   * @param {progressCallbackInfo} args - progress object
+   */
+
+  /**
+   * @typedef {Object} progressCallbackInfo
+   * @property {number} percentage - percentage
+   * @property {number} activeFileSize - total size of the current file
+   * @property {number} activeFileSent - total bytes of the current file transferred
+   * @property {string} currentFile - current file
+   * @property {string} speed - speed
+   * @property {string} elapsedTime - elapsed time
+   */
+
+  /**
+   * @typedef {function()} completedCallback
+   * @callback completedCallback
+   */
+
+  /**
+   * description - Upload or download files from MTP device to local or vice versa
+   *
+   * @param {string} destination
+   * @param {'upload'|'download'} direction
+   * @param {[string]} fileList
+   * @param {string} storageId
+   * @param {errorCallback} onError
+   * @param {progressCallback} onProgress
+   * @param {completedCallback} onCompleted
+   *
+   * @return
+   */
+  transferFiles = ({
+    destination,
+    fileList,
+    direction,
+    storageId,
+    onError,
+    onProgress,
+    onCompleted,
+  }) => {
+    checkIf(onError, 'function');
+    checkIf(onProgress, 'function');
+    checkIf(onCompleted, 'function');
+
+    try {
+      if (undefinedOrNull(destination)) {
+        onError({
+          error: `Invalid path`,
+          stderr: null,
+          data: null,
+        });
+
+        return;
+      }
+
+      const storageSelectCmd = `"storage ${storageId}"`;
+
+      if (isEmpty(fileList) || !isArray(fileList)) {
+        onError({
+          error: `No files selected`,
+          stderr: null,
+          data: null,
+        });
+
+        return;
+      }
+
+      let cmdArgs = [];
+      switch (direction) {
+        case 'download':
+          cmdArgs = (fileList ?? []).map((sourcePath) => {
+            const destinationPath = path.resolve(destination);
+            const escapedDestinationPath = this._escapeShellMtp(
+              `${destinationPath}/${baseName(sourcePath)}`
+            );
+            const escapedSourcePath = `${this._escapeShellMtp(sourcePath)}`;
+
+            return `-e ${storageSelectCmd} "get \\"${escapedSourcePath}\\" \\"${escapedDestinationPath}\\""`;
+          });
+
+          return this._transferFiles({
+            cmdArgs,
+            onError,
+            onProgress,
+            onCompleted,
+          });
+
+        case 'upload':
+          cmdArgs = (fileList ?? []).map((sourcePath) => {
+            const destinationPath = path.resolve(destination);
+            const escapedDestinationPath = `${this._escapeShellMtp(
+              destinationPath
+            )}`;
+            const escapedSourcePath = `${this._escapeShellMtp(sourcePath)}`;
+
+            return `-e ${storageSelectCmd} "put \\"${escapedSourcePath}\\" \\"${escapedDestinationPath}\\""`;
+          });
+
+          return this._transferFiles({
+            cmdArgs,
+            onError,
+            onProgress,
+            onCompleted,
+          });
+
+        default:
+          break;
+      }
+    } catch (e) {
+      log.error(e);
+    }
+  };
 }
